chore: move getLocators to utils file
