+++ app/renderer/actions/Inspector.js
@@ -3,10 +3,10 @@ import {v4 as UUID} from 'uuid';
 
 import i18n from '../../configs/i18next.config.renderer';
 import {SAVED_FRAMEWORK, SET_SAVED_GESTURES, getSetting, setSetting} from '../../shared/settings';
-import {APP_MODE, getLocators} from '../components/Inspector/shared';
+import {APP_MODE} from '../components/Inspector/shared';
 import AppiumClient, {NATIVE_APP} from '../lib/appium-client';
 import frameworks from '../lib/client-frameworks';
-import {findElementByPath, xmlToJSON} from '../util';
+import {findElementByPath, getSimpleSuggestedLocators, xmlToJSON} from '../util';
 import {showError} from './Session';
 
 export const SET_SESSION_DETAILS = 'SET_SESSION_DETAILS';
@@ -149,7 +149,7 @@ export function selectElement(path) {
     dispatch({type: SET_EXPANDED_PATHS, paths: copiedExpandedPaths});
 
     // Calculate the recommended locator strategies
-    const strategyMap = _.toPairs(getLocators(selectedElement.attributes, sourceXML));
+    const strategyMap = _.toPairs(getSimpleSuggestedLocators(selectedElement.attributes, sourceXML));
     dispatch({type: SET_OPTIMAL_LOCATORS, strategyMap});
 
     // Debounce find element so that if another element is selected shortly after, cancel the previous search
